requirements.txt
# FastAPI & Core Web Services
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.8.2
python-multipart==0.0.6
requests>=2.31.0

# Reinforcement Learning & ML
stable-baselines3==2.2.1
stable-baselines3[extra]==2.2.1
gymnasium==0.29.1
shimmy>=0.2.1
torch>=2.2.0

# LLM & Text Embeddings
transformers==4.36.2
sentence-transformers==2.3.1

# General NLP & Text Processing
spacy==3.7.2
nltk==3.8.1
textstat==0.7.3
pymupdf==1.23.8

# ML & Embeddings
transformers==4.36.2
sentence-transformers==2.3.1
torch>=2.2.0
numpy==1.26.3
scikit-learn==1.4.0
joblib==1.3.2

# TTS
edge-tts==6.1.9

# Data Processing
pandas==2.1.4

# Database clients
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
pytest==7.4.4
aiofiles==23.2.1
